refactor: Move demo data to shared presentation-model module
Move DemoData (door event fixtures) from androidApp to
presentation-model/commonMain so iOS previews can reuse the same
test data. Replace java.time.Instant.parse with a Long constant
for KMP compatibility. Add presentation-model dependency to androidApp.
